change amount data type to double

+++ wte-backend/src/main/java/net/summer23project/wtebackend/entity/UserIngredientInventory.java
@@ -24,7 +24,7 @@ public class UserIngredientInventory {
     private Long id;
 
     @Column(name = "ingredient_amount", nullable = false)
-    private int ingredientAmount;
+    private double ingredientAmount;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
